Add explicit option to use shared memory and fixes bug when input list is empty (#120)
* Add flag to set shmem to off by default

* Change workflow script to use shmem flag

* Add fallback message

* Dealt with empty list

* Should actually fix empty list problem now

+++ umbridge/um.py
@@ -51,7 +51,7 @@ def supported_models(url):
   return response["models"]
 
 class HTTPModel(Model):
-    def __init__(self, url, name):
+    def __init__(self, url, name, use_shmem=False):
         super().__init__(name)
         self.url = url
 
@@ -67,29 +67,30 @@ class HTTPModel(Model):
         self.__supports_apply_hessian = response["support"].get("ApplyHessian", False)
         self.__supports_shmem = False
 
-        #Test whether client and server are able to communicate through shared memory. Disables ShMem if test fails.
-        testvec = [12345.0]
-        tid = threading.get_native_id()
-        input["tid"] = str(tid)
-        shm_c_in = shared_memory.SharedMemory("/umbridge_test_shmem_in_" + str(tid), True, 8)
-        raw_shmem_input = np.ndarray(1, dtype=np.float64, buffer=shm_c_in.buf)
-        raw_shmem_input[:] = testvec[0]
-        shm_c_out = shared_memory.SharedMemory("/umbridge_test_shmem_out_" + str(tid), create=True, size=8)
-        raw_shmem_output = np.ndarray(1, dtype=np.float64, buffer=shm_c_out.buf)
-        try: response = requests.post(f"{self.url}/TestShMem", json=input).json()
-        except: pass
-        result = []
-        result.append(raw_shmem_output.tolist()[0])
-        shm_c_in.close()
-        shm_c_in.unlink()
-        shm_c_out.close()
-        shm_c_out.unlink()
+        if use_shmem:
+            #Test whether client and server are able to communicate through shared memory. Disables ShMem if test fails.
+            testvec = [12345.0]
+            tid = threading.get_native_id()
+            input["tid"] = str(tid)
+            shm_c_in = shared_memory.SharedMemory("/umbridge_test_shmem_in_" + str(tid), True, 8)
+            raw_shmem_input = np.ndarray(1, dtype=np.float64, buffer=shm_c_in.buf)
+            raw_shmem_input[:] = testvec[0]
+            shm_c_out = shared_memory.SharedMemory("/umbridge_test_shmem_out_" + str(tid), create=True, size=8)
+            raw_shmem_output = np.ndarray(1, dtype=np.float64, buffer=shm_c_out.buf)
+            try: response = requests.post(f"{self.url}/TestShMem", json=input).json()
+            except: pass
+            result = []
+            result.append(raw_shmem_output.tolist()[0])
+            shm_c_in.close()
+            shm_c_in.unlink()
+            shm_c_out.close()
+            shm_c_out.unlink()
 
-        if(result[0] != testvec[0]):
-            print("Server not accessible via shared memory")
-        else:
-            self.__supports_shmem= True
-            print("Server accessible via shared memory")
+            if(result[0] != testvec[0]):
+                print("Server not accessible via shared memory. Using HTTP instead.")
+            else:
+                self.__supports_shmem= True
+                print("Server accessible via shared memory")
 
 
     def get_input_sizes(self, config={}):
@@ -121,7 +122,7 @@ class HTTPModel(Model):
     def supports_shmem(self):
         return self.__supports_shmem
     
-    def __check_input_is_list_of_lists(self,parameters):
+    def __check_input_is_list_of_lists(self, parameters):
         if not isinstance(parameters, list):
             raise Exception("Parameters must be a list of lists!")
         if not all(isinstance(x, list) for x in parameters):
@@ -143,10 +144,14 @@ class HTTPModel(Model):
 
             for i in range(len(parameters)):
                 inputParams["shmem_size_" + str(i)] = len(parameters[i])
-                shm_c_in = shared_memory.SharedMemory(inputParams["shmem_name"] + "_in_" + str(tid) + f"_{i}", create=True, size=len(parameters[i])*8)
-                raw_shmem_input = np.ndarray((len(parameters[i]),), dtype=np.float64, buffer=shm_c_in.buf)
-                raw_shmem_input[:] = parameters[i]
-                buffers.append(shm_c_in)
+                if len(parameters[i]) > 0:
+                    shm_c_in = shared_memory.SharedMemory(inputParams["shmem_name"] + "_in_" + str(tid) + f"_{i}", create=True, size=len(parameters[i])*8)
+                    raw_shmem_input = np.ndarray((len(parameters[i]),), dtype=np.float64, buffer=shm_c_in.buf)
+                    raw_shmem_input[:] = parameters[i]
+                    buffers.append(shm_c_in)
+                else: # Handles edge case with empty list
+                    continue
+                
             output_sizes = self.get_output_sizes(config)
 
             for i in range(len(output_sizes)):
@@ -442,10 +447,13 @@ def serve_models(models, port=4242, max_workers=1, error_checks=True):
 
         parameters = []
         for i in range(req_json["shmem_num_inputs"]):
-            shm_c_in = shared_memory.SharedMemory(req_json["shmem_name"] + "_in_" + str(req_json["tid"]) + f"_{i}", False, req_json[f"shmem_size_{i}"])
-            raw_shmem_parameter = np.ndarray((req_json[f"shmem_size_{i}"],), dtype=np.float64, buffer=shm_c_in.buf)
-            parameters.append(raw_shmem_parameter.tolist())
-            shm_c_in.close()
+            if req_json[f"shmem_size_{i}"] == 0: # Handles edge case with empty list
+                parameters.append([])
+            else:
+                shm_c_in = shared_memory.SharedMemory(req_json["shmem_name"] + "_in_" + str(req_json["tid"]) + f"_{i}", False, req_json[f"shmem_size_{i}"])
+                raw_shmem_parameter = np.ndarray((req_json[f"shmem_size_{i}"],), dtype=np.float64, buffer=shm_c_in.buf)
+                parameters.append(raw_shmem_parameter.tolist())
+                shm_c_in.close()
 
         # Check if parameter dimensions match model input sizes
         if len(parameters) != len(model.get_input_sizes(config)):
